Removed compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/irrlicht@4081 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/guiengine/modaldialog.cpp
@@ -43,8 +43,8 @@ ModalDialog::ModalDialog(const float percentWidth, const float percentHeight)
     
     assert(w > 0);
     assert(h > 0);
-    assert(w <= frame_size.Width);
-    assert(h <= frame_size.Height);
+    assert((unsigned int)w <= frame_size.Width);
+    assert((unsigned int)h <= frame_size.Height);
     
     m_area = core::rect< s32 >( position2d< s32 >(frame_size.Width/2 - w/2, frame_size.Height/2 - h/2),
                                dimension2d< s32 >(w, h) );
